Don't append extra NUL to URLs in wxGTK wxURLDataObject
The buffer size includes the trailing NUL, but we shouldn't make it part
of the string contents as it's not really part of the data at all.

The next commit will add a test verifying that retrieving an URL from
clipboard now gets the right string, without an extra trailing NUL, i.e.
fixes a bug similar to #22928 but in wxGTK.

As with the grandparent commit, this commit is a modified version of the
master one accepting non-NUL-terminated strings too for compatibility.

See #23018.

(cherry picked from commit c93c346adcd0f7cb443f6e8307aed6ccb59637af)