Add wxUxThemeHandle::GetColour() replacing MSWGetThemeColour()
Add a helper function wrapping ::GetThemeColor() and use it in all
places where this function was used before, as it's more convenient.

Also remove wxWindow::MSWGetThemeColour() as it was only used in a
single place and doesn't seem to present any advantages compared to
using wxUxThemeHandle new method.

There should be no changes in behaviour, this is just a refactoring.