Fix scaling of standard MSW platform icons in high DPI
Use the correct scaling factor when creating the bitmap in
wxWindowsArtProvider, otherwise the bitmap was wrongly rescaled again
when actually using it.

It still seems wrong to always use the primary display scaling here, but
we just don't have any window to use in this code and at least it is
consistent and works when primary display scaling is not 100% now.

Closes #23031.