Allow selecting and copying text in wxMessageDialog on GTK
Normally the GTK message dialogs have non-selectable text, but that also
means users can't copy the text. This can be inconvenient if the dialog
is used for error messages, since then the error message can't be
copied.

Instead, make the text labels used in the GTK message dialog selectable
to fix that.

Closes #23039.