Sort only by this column in wxDataViewColumn::SetSortOrder()
Make the generic version behave consistently with the native ones and
replace any existing sort columns with the current one unless multi
column sort is allowed (which is not the case by default and not
supported by the native versions).

This ensures that calling SetSortOrder() from the program now has the
same effect as clicking on the column header interactively.

This change is not backwards-compatible but can be arguably considered
as a bug fix, as it doesn't seem like the old behaviour, which showed
only one sort column in the UI, but sorted using multiple columns
internally, could be ever useful.