Reset old sort indicator in GTK wxDataViewColumn::SetSortOrder()
It doesn't make sense to show sort indicator for more than one column in
the current GTK wxDataViewCtrl implementation because it always uses
exactly one column for the actual sorting, but this is exactly what
happened if SetSortOrder() was called on a column different from the
currently used one.

Explicitly reset the sort indicator shown for the old sort column to
prevent this from happening.