wxOverlay implementation for wxMSW
Fixes issue #23047 wxOverlay broken by use of WS_EX_COMPOSITED
because we can no longer draw outside the paint event when double
buffering is enabled. i.e. using wxClientDC to draw a rubberband
won't have any effect on a window with WS_EX_COMPOSITED flag on and
the drawing will be simply lost. Nevertheless, the drawing can still
be accomplished by hijacking the associated dc to draw on an offscreen
bitmap, then applying it to the overlay window ( a toplevel window
created with WS_EX_LAYERED ) when drawing is complete.