Use wxRadioButton in MSW wxRadioBox implementation
For reasons lost in the mists of time, wxRadioBox created its own radio
buttons using Win32 API directly instead of just using wxRadioButton.

Finally change this now as this significantly simplifies the code and
makes it shorter and more clear. Notably, we now can just add the
buttons to the sizer and let it manage them instead of having to do it
manually.

It also makes an ugly wxRadioBox-specific workaround in wxWindow
unnecessary. Unfortunately we still need to keep wxSubwindows stuff
because wxSlider still uses it, but this will make it easier to get rid
of it in the future, when wxSlider is changed to use wxWindows too.

This will also make the radio box appear correctly in the dark mode too,
as a nice side effect.