Fix memory leak of command line arguments in wxMSW
Use the same hack already used for wxEntryReal() before to define a
MSW-specific wxEntryCleanup() which frees the memory allocated by
::CommandLineToArgvW() after performing the common cleanup.

Also refactor the code dealing with them to avoid having different
wxArgs variables in wxBase and wxCode, which was almost certainly
unintentional and definitely very confusing. Now we keep just a single
wxArgs in wxBase and provide access to it from wxCore via the exported
functions.

Closes #23081.