Set appropriate colours for wxStyledTextCtrl in dark mode
Note that the applications still need to be adapted to use different
colours for the dark mode too, as the control can't set the colours for
all the styles used by various lexers.

But at the very least this sets good defaults and it's now possible to
use StyleGet{Fore,Back}ground() to construct other colours based on
them.

See #19107.