Never call wxMenu::HandleCommandProcess() with null item
This probably didn't happen even before, but it wasn't really clear as
some code checked for it and some didn't. Now check for it everywhere
(except in the places where this pointer is already dereferenced
unconditionally before calling HandleCommandProcess()) and also clearly
specify that it's a precondition for calling it inside the function
itself.

No real changes.