Replace macros with template function in wxString::ToXXX()
Get rid of the ugly non-hygienic macros and use a template function
instead.

The code is still not as simple as it would ideally be because we need
to use lambdas in a couple of places instead of passing the conversion
function directly and then to explicitly specify the template argument
because it can't be deduced from lambda, but is more clear and easier to
debug.

No real changes.