Use correct size for SVG art provider bitmaps in high DPI
The base size passed to wxBitmapBundle::FromSVG() must _not_ depend on
the current DPI as it specifies the size without scaling, but we passed
it the DPI-dependent size by using GetSizeHint() if hte desired size
wasn't specified.

Fix this now just by using GetDIPSizeHint() instead.

This notably fixes the sizes of the generic (i.e. not provided by
wxWindowsArtProvider) icons in the print preview frame when the main
monitor uses high DPI.