Fix painting artifacts with AUI and wxGTK3 in some environments
Using wxClientDC can actually be worse than useless with wxGTK3 as not
only it does nothing when used under Wayland, but it may also draw at
the wrong position when using X11 with some WMs, notably Openbox or
fvwm.

Simply don't use at all with wxGTK3, it should be never needed there, to
fix the painting artifacts that it created.

Closes #23135.