Make wx/debug.h included as first wx header compile again
This was broken by 0f6c54cdb6 (Define __WXFUNCTION__ as __func__ and
don't use it any more, 2022-11-11) which mistakenly removed the
inclusion of wx/cpp.h from here under the belief that it was only needed
for __WXFUNCTION__, as its comment indicated.

But this isn't really the case, so restore the removed #include and just
update its comment.