Replace wxListItemData::Init() with member initialization
Get rid of Init() function and just initialize all the members in their
declarations, as this is less verbose, more clear and more robust.

Note that we also initialize m_rect now, which was left uninitialized by
Init() before, as it's simpler and more reliable to always do even if we
happen to overwrite it in the ctor.