Replace wxVector<wxIdRange*> with std::vector<wxIdRange>
Don't use heap-allocated pointers needlessly, when we can just use the
values directly.

Also remove the useless Find() and FindRangeForItem() functions as it's
just as simple to iterate over all items than to call them.

We may want to use a map instead of a vector here later, but for now
this is already a simplification (and a micro optimization).