Make simple MSW About dialog more consistent with the generic one
Show short version on the same line as the program name and put the
copyright at the bottom, as in the other versions.

This makes the simple message box more similar to the fuller dialog
shown by wxAboutBox() if the web site or any other "non simple" field is
specified.