Fix return value of wxToolBar::GetToolBitmapSize() under GTK/Mac
Fix another pixel units confusion in wxToolBar code and ensure that
"sizeNeeded" is expressed in logical units when a fixed bitmap size is
set, to make it consistent with the case when it is not set, and also
because bitmap size is supposed to be expressed in logical pixels, in
both SetToolBitmapSize() and, for consistency, GetToolBitmapSize() too.

Also explicitly mention the latter in the documentation.

See #23228, #23222.

(cherry picked from commit 453a52b942fa27c8f412716c8290dbcb6172d0e4)