Fix wxStaticBoxSizer with sibling windows in nested sizers too
This extends the workaround implemented in the previous commits to deal
with the case of sibling windows being added to wxStaticBoxSizer not
directly but as children of a nested sizer.

To handle this case, we need to remember if we have any nested sizers
and examine them recursively for non-box children. To avoid having to
make this, relatively expansive, check on each layout, remember its
result after the first layout. This is not infallible, e.g. it doesn't
handle the scenario in which a sibling window is added to a nested sizer
later, but this seems to be the best we can do without too much overhead
for the normal case and such cases should hopefully be very uncommon.

Also note that we now only give the debug message about the sizer item
having the wrong parent only once instead of doing it for each item, but
this might actually be better, as it's not clear that getting many
almost identical messages for the sizers containing many items is really
more useful.

This commit is best viewed using Git --color-moved option.