Remove useless assert from the sizer code
Asserting that a pointer is non-null before dereferencing it is useless
as if it were null, we'd crash anyhow.

In this particular case, the pointer is guaranteed to never be null
because we only call Item() if the index passed to it is valid, so the
assert can never be triggered.