Ensure wxRadioBox is big enough to fit its contents in wxMSW
Since the changes of ce1ee46b0c (Use wxRadioButton in MSW wxRadioBox
implementation, 2022-12-24) the control could be too small because we
only accounted for the extra static box margins in two directions (top
and left) but not the two other ones.

Fix this by adding the border around the buttons in all directions and
also increasing its size to preserve more or less the same appearance as
before.

Closes #23267.