Apply workaround to all sibling children, not just the first one
This is important as while giving debug message is optional (and might
even be not ideal, see the grandparent commit description), we do need
to call MSWDisableComposited() when all sibling children already exist
as otherwise it wouldn't reset WS_EX_COMPOSITED for them -- and many
windows, such as wxOwnerDrawnComboBox are not drawn correctly if they
have WS_EX_COMPOSITED and overlap a sibling window (worse, this somehow
seems to break the redrawing of the entire TLW, with even the controls
completely outside of the static box being affected, see #23239).

This commit is best viewed ignoring whitespace-only changes.