Use custom painting in wxStaticBox even with double buffering on
Unfortunately the fact that is on when wxStaticBox is created doesn't
guarantee that it remains on when it is painted later, as
MSWDisableComposited() could be called in the meanwhile, breaking the
redraw of the box.

Conservatively fix this by just always painting the box ourselves, but
mention a possibly better solution.