Draw borders for scrollable windows in wxGTK too
In many themes tree view window doesn't have a border, and so none of
our windows using wxHSCROLL and/or wxVSCROLL styles had it neither, even
if they didn't have wxNO_BORDER style.

Fix this by using the same border as is used for the entry widget (which
usually does have it) for all windows, not just those without scroll
styles.

In particular, this fixes the problem with the missing border for
wxListCtrl (which uses scrolling).

Closes #23236.