Fix installing httpbin in Ubuntu 18.04 CI builds
For some reason the wheel (i.e. binary package) that was installed by
python3-pip 9.0.1-2.3~ubuntu1.18.04.6 before is not installed by the
latest 9.0.1-2.3~ubuntu1.18.04.7 version of it which has just appeared
in Ubuntu 18.04 repositories, and cffi is now built as part of the
module installation, see the following pseudo-diff between the old
successful and the new failing builds:

- Downloading https://files.pythonhosted.org/packages/3a/12/d6066828014b9ccb2bbb8e1d9dc28872d20669b65aeb4a86806a0757813f/cffi-1.15.1-cp36-cp36m-manylinux_2_5_x86_64.manylinux1_x86_64.whl
+ Downloading https://files.pythonhosted.org/packages/2b/a8/050ab4f0c3d4c1b8aaa805f70e26e84d0e27004907c5b8ecc1d31815f92a/cffi-1.15.1.tar.gz

And building this module requires libffi, so install it.