Change container type for wxAuiToolBar's item storage
Change the container type used by wxAuiToolBar to be a list of pointers
instead of an array of items. This allows wxAuiToolBarItem to be
subclassed, and then for those subclasses to be used in the toolbar.

This also allows for modernizing the code to use range-based for loops,
which weren't supported by the previous array implementation.