Introduce wxAuiToolBarItemKind to replace AUI use of wxItemKind
Not all item kinds for wxAuiToolBarItem were in wxItemKind (or exported
to the user-facing API), so introduce separate AUI toolbar item kinds to
fully describe wxAuiToolBarItem kinds.

The API for wxAuiToolBar::AddTool was loosend to be an int to allow using
either the legacy wxItemKind values or the new wxAuiToolBarItemKind
values as an argument.