Use better behaving custom validator in wxDatePickerCtrlGeneric
Using wxTextValidator with wxFILTER_INCLUDE_CHAR_LIST was both too
restrictive, as it prevented the user from entering any letters, even if
the locale format required using them for months (as e.g. nb-NO does)
and too lax, as it allowed entering any invalid date as long as it
contained only valid characters.

Fix both of these problems by using a custom validator actually using
wxDateTime::ParseFormat() for validation instead of just checking that
the string consists of valid characters (and doing it wrongly).

This validator could be more generally useful, and so could be extracted
into a public class later, but for now use it only here.

Closes #23310.