Improve workaround for wrong wxAppTraits creation
Instead of creating wrong wxAppTraits before wxApp object is fully
constructed, don't create them at all by checking whether the object is
complete in GetTraits().

This is better than the version of the grandparent commit because it
doesn't hide accidental traits creation, as was e.g. the case in wxMSW
wxApp until the change of the parent commit, any longer.

See #23315.