Move thread-specific info cleanup to wxThread::Exit()
This function was somehow being called too late, at least according to
TSAN, and resulted in a data race on the global wxAllThreadInfos vector
because the main thread could destroy it before the worker thread
exited, apparently.

It's not really clear how it happened but it seems that the dtor of a
local object in PthreadStart was called at unexpected moment, so don't
rely on RAII scope guard and just call ThreadCleanUp() manually right
before exiting the thread in Exit().

This mechanism is still fragile, not to say broken, e.g. we don't seem
to clean up thread-specific info at all if the thread is killed, but at
least we don't get TSAN data race warnings whenever a detached thread is
destroyed now.