Fix data race in wxGetTimeZone()
We could return a wrong value from this function if two threads called
it simultaneously, as gmtoffset value could be returned without DST
adjustment in one of them. And just accessing it from two different
threads without synchronization was a data race on its own.

Fix both problems by using an atomic int for it and taking care to only
set it to the correct value.