Draw the border at the same vertical position even without label
This looks better as otherwise we are left with a too big gap between
the top of the control and the contents, while like this we draw the
top border line at the same position as for the boxes using a label.

Alternative solution, that would result in a more compact layout, would
be to reduce the vertical border used by wxStaticBox when it doesn't
have any label, but this would make boxes with and without label look
very different, which is arguably not desirable.