Fix on wxPopupWindow to allow full use of focusable controls on GTK+
Presently on GTK+ a POPUP_WINDOW is created when instantiating the
popup; but having focusable controls on it is not supported by GTK+,
as evidenced by the sample.
On the sample the text control on it can be modified by keyboard
input but will not show a caret while being modified, due to a GTK
limitation.
This fix instead creates a TOPLEVEL_WINDOW and hides it from the
pager and window bar, effectively allowing the focus to now be
represented with a caret on text controls.
Since the MSW port already had a style to support controls on the
popup this fix uses the same on GTK+, with the same meaning.
During testing it was shown that a TAB press would force the popup
to center on the screen so code was added to reposition the popup
when needed.