Fix wxPropertyGrid tab traversal under wxGTK
Tab traversal from a wxPropertyGrid didn't work across wxPanel
instances or when the property grid was followed by a static control
such as wxStaticText due to implementing it "manually".

Fix this by using Navigate() to implement this instead, and just keep a
SetFocus() hack which is still needed for wxGTK.

Fixes #23354.

Closes #23358.