Document wxGLCanvas::CreateSurface()
It is needed to be able to re-create the EGL drawing surface after
the window layout have changed, such as after a reparent of the canvas
or of it's grandparents.

The functionality is already there, so this is mostly a documentation
to keep the functionality available in future versions.

Also destroy the surface before re-creating the surface, so the user
don't have to do that.